Refactor: split SpellChecker core (libuilogic) from interactive SpellCheckManager
Extract the framework-agnostic spell-check core (Hunspell load, names/user/skip/
change-all lists, word classification, Initialize/IsWordCorrect/GetSuggestions)
into a new SpellChecker base class in libuilogic. SpellCheckManager (UI) now
inherits it and keeps only the interactive, SubtitleLineViewModel-bound flow
(CheckSpelling/ChangeWord). Moves the WeCantSpell.Hunspell package ref to
libuilogic. seconv can now instantiate SpellChecker directly as an ISpellChecker.
